get literalWidget generator from vdom config
