+++ lib/compiler.js
@@ -31,6 +31,7 @@ function Compiler(node, options) {
 
   // get vdom options
   this.vdom = VDOM_CONFIG[options.vdom || `virtual-dom`];
+  this.generateLiteral = this.vdom.generateLiteral;
   this.marshalDataset = this.vdom.dataset;
   this.rawProps = this.vdom.rawProps;
 
@@ -91,39 +92,7 @@ Compiler.prototype.compile = function() {
   }
 
   if (this.hasLiteral) {
-    js = `
-      function generateLiteralWidget(id, contents) {
-        function LiteralWidget(id, contents) {
-          this.name = 'LiteralWidget'
-          this.id = id
-          this.contents = contents
-        }
-        LiteralWidget.prototype.type = 'Widget'
-        LiteralWidget.prototype.init = function () {
-          var wrapper = document.createElement('div')
-          wrapper.innerHTML = this.contents
-          var root
-          if (wrapper.childNodes.length === 1) {
-            root = wrapper.firstChild
-          } else {
-            root = wrapper
-          }
-          return root
-        }
-        LiteralWidget.prototype.update = function (previous, domNode) {
-          return domNode
-        }
-        // 'render' is called by the vdom-to-html module which is used in the unit tests
-        LiteralWidget.prototype.render = function () {
-          var h = require('virtual-dom/h')
-          var host = document.createElement('div')
-          host.appendChild(this.init())
-          return h('text', host.innerHTML)
-        }
-        return new LiteralWidget(id, contents)
-      };
-      ${js}
-    `;
+    js = `${this.generateLiteral}${js}`;
   }
 
   return js;
